requirements.txt
flask==3.0.0
pymongo==4.5.0
redis==4.5.5
neo4j==5.11.0
python-dotenv==1.0.0